Shift over the other class (but not yet tested)
