+++ guidance/models/_azureai.py
@@ -1,7 +1,19 @@
-from typing import Callable, Optional
+from typing import Callable, Iterator, Optional, Union
 
+from pydantic import TypeAdapter
+
+from .._ast import (
+    JsonNode,
+    RuleNode,
+)
 from ._base import Model
-from ._openai_base import BaseOpenAIInterpreter, OpenAIAudioMixin, OpenAIImageMixin
+from ._openai_base import (
+    BaseOpenAIInterpreter,
+    OpenAIAudioMixin,
+    OpenAIImageMixin,
+    Message,
+)
+from ..trace import OutputAttr
 
 
 class AzureOpenAIInterpreter(BaseOpenAIInterpreter):
@@ -96,3 +108,92 @@ def create_azure_openai_model(
     model = Model(interpreter=interpreter, echo=echo)
 
     return model
+
+
+class AzureInferenceInterpreter(BaseOpenAIInterpreter):
+    def __init__(
+        self,
+        *,
+        endpoint: str,
+        credential: Union["AzureKeyCredential", "TokenCredential"],
+    ):
+        try:
+            import azure.ai.inference
+        except ImportError:
+            raise Exception(
+                "Please install the azure-ai-inference package  using `pip install azure-ai-inference` in order to use guidance.models.AzureInference!"
+            )
+        client = azure.ai.inference.ChatCompletionsClient(
+            endpoint=endpoint,
+            credential=credential,
+        )
+        super().__init__(client=client)
+
+    def _run(self, **kwargs) -> Iterator[OutputAttr]:
+        if self.state.active_role is None:
+            # Should never happen?
+            raise ValueError(
+                "OpenAI models require chat blocks (e.g. use `with assistant(): ...`)"
+            )
+        if self.state.active_role != "assistant":
+            raise ValueError(
+                "OpenAI models can only generate as the assistant (i.e. inside of `with assistant(): ...`)"
+            )
+        if self.state.content:
+            raise ValueError(
+                f"OpenAI models do not support pre-filled assistant messages: got data {self.state.content}."
+            )
+
+        with self.client.complete(
+            body={
+                "messages": TypeAdapter(list[Message]).dump_python(self.state.messages),
+                "log_probs": self.log_probs,
+                "stream": True,
+                **kwargs,
+            },
+            headers={
+                "extra-parameters": "pass-through",
+            },
+        ) as chunks:
+            yield from self._handle_stream(chunks)
+
+    def rule(self, node: RuleNode, **kwargs) -> Iterator[OutputAttr]:
+        raise ValueError("Rule nodes are not supported for Azure Inference")
+
+    def json(self, node: JsonNode, **kwargs) -> Iterator[OutputAttr]:
+        return self._run(
+            json_schema={
+                "name": "json_schema",  # TODO?
+                "schema": node.schema,
+                "strict": True,
+            },
+            **kwargs,
+        )
+
+
+class AzureInference(Model):
+    def __init__(
+        self,
+        *,
+        endpoint: str,
+        credential: Union["AzureKeyCredential", "TokenCredential"],
+        echo: bool = True,
+    ):
+        """Build a new Azure Inference model object that represents a model in a given state.
+
+        Parameters
+        ----------
+        endpoint : str
+            The endpoint of the Azure Inference service.
+        credential : AzureKeyCredential | TokenCredential
+            The credential to use for authentication with the Azure Inference service.
+        echo : bool
+            If true the final result of creating this model state will be displayed (as HTML in a notebook).
+        """
+        super().__init__(
+            interpreter=AzureInferenceInterpreter(
+                endpoint=endpoint,
+                credential=credential,
+            ),
+            echo=echo,
+        )
